Minor refactoring within hand-made properties
