+++ src/main/java/net/jqwik/properties/arbitraries/RandomGenerator.java
@@ -1,0 +1,8 @@
+package net.jqwik.properties.arbitraries;
+
+import java.util.*;
+
+public interface RandomGenerator<T> {
+
+	T next(Random random);
+}
